[SOK-27] fix: Настроил fps игры
