+++ packages/client/src/components/Game/enemy.tsx
@@ -1,9 +1,9 @@
 import { getRandomEdgePosition } from './utils'
-import { Enemy, Obstacle, Player } from '@/components/Game/gameTypes'
+import { Enemy, Player } from '@/components/Game/gameTypes'
 
-export const initializeEnemies = (): Enemy[] => {
+export const initializeEnemies = (numberOfEnemies: number): Enemy[] => {
   const initialEnemies: Enemy[] = []
-  for (let i = 0; i < 5; i++) {
+  for (let i = 0; i < numberOfEnemies; i++) {
     // количество врагов
     const { x, y } = getRandomEdgePosition(800, 600)
     const enemy: Enemy = {
@@ -22,8 +22,7 @@ export const initializeEnemies = (): Enemy[] => {
 export const updateEnemyPositions = (
   player: Player,
   enemies: Enemy[],
-  setEnemies: React.Dispatch<React.SetStateAction<Enemy[]>>,
-  speedFactor: number
+  setEnemies: React.Dispatch<React.SetStateAction<Enemy[]>>
 ) => {
   setEnemies(prevEnemies =>
     prevEnemies.map(enemy => {
@@ -34,28 +33,16 @@ export const updateEnemyPositions = (
       const normalizedX = directionX / magnitude
       const normalizedY = directionY / magnitude
 
-      const newX = enemy.x + normalizedX * enemy.speed * speedFactor
-      const newY = enemy.y + normalizedY * enemy.speed * speedFactor
+      const newX = enemy.x + normalizedX * enemy.speed
+      const newY = enemy.y + normalizedY * enemy.speed
 
       return { ...enemy, x: newX, y: newY }
     })
   )
 }
 
-export const detectEnemyCollision = (
-  rect1: Player | Enemy,
-  rect2: Obstacle | Enemy
-): boolean => {
-  return (
-    rect1.x < rect2.x + rect2.width &&
-    rect1.x + rect1.width > rect2.x &&
-    rect1.y < rect2.y + rect2.height &&
-    rect1.y + rect1.height > rect2.y
-  )
-}
-
 export const respawnEnemies = (
   setEnemies: React.Dispatch<React.SetStateAction<Enemy[]>>
 ) => {
-  setEnemies(initializeEnemies())
+  setEnemies(initializeEnemies(5))
 }
