+++ packages/client/src/components/Game/gameTypes.tsx
@@ -30,7 +30,6 @@ export interface KeyMap {
 export interface ControlsProps {
   player: Player
   setPlayer: React.Dispatch<React.SetStateAction<Player>>
-  speedFactor: number
   obstacles: Obstacle[]
   canvasWidth: number
   canvasHeight: number
